% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3d.R
\name{ellipse}
\alias{ellipse}
\title{Get ellipse}
\usage{
ellipse(x = 0, y = 0, major = 1, minor = 1, theta = 0, nv = 10)
}
\arguments{
\item{x}{start position x, defaults 0}

\item{y}{start position y, defaults 0}

\item{major}{radius major axis, defaults 1}

\item{minor}{radius minor axis, defaults 1}

\item{theta}{rotaiton angle, defaults 0}

\item{nv}{number of z points, defaults 10}
}
\description{
Get ellipse to build Plotly 3D plot of shape information
}
\author{
Sven Gastauer
}
