% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3d.R
\name{get_shp3d}
\alias{get_shp3d}
\title{Get 3d shape plot}
\usage{
get_shp3d(
  shape,
  name = "",
  domain = list(x = c(0, 1), y = c(0, 1)),
  scene = FALSE,
  out = FALSE
)
}
\arguments{
\item{shape}{list or dataframe containing the shape information}
}
\description{
Plotly 3D plot of shape information
}
\examples{
get_shp3d(list(data(pil_fb), data(pil_sb)))
}
\author{
Sven Gastauer
}
