% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KRM_new.R
\name{krm.fluid}
\alias{krm.fluid}
\title{Fluid model  - KRM
Based on equations found in 'Acoustic models of fish: The Atlantic cod (Gadus morhua)' by Clay and Horne, 1994}
\usage{
krm.fluid(
  frequency = 120000,
  c.w = 1490,
  c.b = 1570,
  rho.w = 1030,
  rho.b = 1070,
  x,
  w,
  z_U,
  z_L,
  theta = 90
)
}
\arguments{
\item{frequency}{Frequency in Hz defaults to 120000}

\item{c.w}{ambient soundspeed in m/s defaults to 1490}

\item{c.b}{Soundspeed in fish body m/s defautls to 1570}

\item{rho.w}{Ambient water Density kg/m3 defaults to 1030}

\item{rho.b}{Density fish body kg/m3 defaults to 1070}

\item{x}{Position x direction Fish body}

\item{w}{Width at Position x direction Fish body}

\item{z_U}{Height fish body at position x, upper value}

\item{z_L}{Height fish body at position x, lower value}

\item{theta}{incident angle in degrees, note that 90 degrees equals broadside incident, KRM is only stable for values between 65 and 115 degrees, defaults 90 degrees}
}
\description{
Run fluid KRM model
}
\examples{
data(pil_fb)
f = krm.fluid(frequency=120000, c.w=1490, c.b=1570, x=pil_fb$x_fb, w=pil_fb$w_fb, z_U=pil_fb$z_fbU, z_L = pil_fb$z_fbL, theta=90)
sigma=abs(f)
TS=20 * log10(sigma)
}
\author{
Sven Gastauer
}
