% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KRM_new.R
\name{krm}
\alias{krm}
\title{Run KRM model
Based on equations found in 'Acoustic models of fish: The Atlantic cod (Gadus morhua)' by Clay and Horne, 1994}
\usage{
krm(
  frequency = 120 * 1000,
  c.w = 1490,
  rho.w = 1030,
  theta = 90,
  cs = c(1570, 345),
  rhos = c(1070, 1.24),
  shape = list(pil_fb, pil_sb),
  modes = c("fluid", "soft"),
  L = 0.21,
  fb = 1
)
}
\arguments{
\item{frequency}{Frequency in Hz defaults to 120000}

\item{c.w}{ambient soundspeed in m/s defaults to 1490}

\item{rho.w}{Ambient water Density kg/m3 defaults to 1030}

\item{theta}{incident angle in degrees, note that 90 degrees equals broadside incident, KRM is only stable for values between 65 and 115 degrees, defaults 90 degrees}

\item{cs}{Soundspeeds in m/s for the different body parts, defautls to c(1570, 345), typical for fishbody and swimbladder}

\item{rhos}{Densities in kg/m3 for the different body parts, defaults to c(1070,345) , typical for fishbody and swimbladder}

\item{shape}{list of shapes that describe the body parts. must contain columns that start with x,w, z (z****L,z****U), describing the position along the x axis, the width, the upper and lower height.}

\item{modes}{which model mode should be used fluid or soft, fluid for weakly scattering nody parts and soft for stronger scattering parts, defaults to c("fluid", "soft"), typical for fish body and swimbladder}

\item{L}{Lengths to be computed in m, defaults to 0.21}

\item{fb}{Which body part is the largest, to which body part should the length be adjusted, defaults to 1 (first body part will be he largest)}
}
\description{
Run KRM model
}
\examples{
data(pil_fb)
data(pil_sb)
TS = krm(frequency=120000, c.w=1490, rho.w=1030, theta=90, cs=c(1570, 345), rhos=c(1070,1.24), shape=list(pil_fb, pil_sb), modes=c("fluid", "soft"), L=0.21, fb=1)
}
\author{
Sven Gastauer
}
