% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KRM_new.R
\name{krm.soft}
\alias{krm.soft}
\title{Soft model  - KRM
Based on equations found in 'Acoustic models of fish: The Atlantic cod (Gadus morhua)' by Clay and Horne, 1994}
\usage{
krm.soft(
  frequency = 120000,
  theta = 90,
  c.w = 1490,
  c.f = NULL,
  c.b = 345,
  rho.w = 1030,
  rho.f = NULL,
  rho.b = 1.24,
  x,
  w,
  z_U
)
}
\arguments{
\item{frequency}{Frequency in Hz defaults to 120000}

\item{theta}{incident angle in degrees, note that 90 degrees equals broadside incident, KRM is only stable for values between 65 and 115 degrees, defaults 90 degrees}

\item{c.w}{ambient soundspeed in m/s defaults to 1490}

\item{c.f}{Soundspeed in fish body m/s defautls to NULL (if NULL, ambient water c.w will be used)}

\item{c.b}{Soundspeed in swimbladder m/s defaults to 345}

\item{rho.w}{Ambient water Density kg/m3 defaults to 1030}

\item{rho.f}{Density fish body kg/m3 defaults to NULL  (if NULL, ambient water rho.w will be used)}

\item{rho.b}{Density gas bladder kg/m3 defaults to 1.24}

\item{x}{Position x direction Fish body}

\item{w}{Width at Position x direction Fish body}

\item{z_U}{Height at position x, upper value}

\item{z_L}{Heightat position x, lower value}
}
\description{
Run soft KRM model
}
\examples{
data(pil_sb)
f = krm.soft(frequency=120000, theta=90, c.w=1490, c.f=NULL, c.b=345, rho.w=1030, rho.f=NULL, rho.b=1.24, x=pil_sb$x_sb, w=pil_sb$w_sb, z_U=pil_sb$z_sbL)
sigma = abs(f)
TS = 20 * log10(sigma)
}
\author{
Sven Gastauer
}
