% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KRM_new.R, R/bessel.R
\name{lambda}
\alias{lambda}
\title{Wavelength Lambda}
\usage{
lambda(c, f)

lambda(c, f)
}
\arguments{
\item{c}{Soundspeed in m/s}

\item{f}{frequency in Hz (s^-1)}

\item{frequency}{Hz}
}
\value{
lambda wavelength m
}
\description{
Wavelength Lambda

Calculate the acoustic wave length lambda
}
\examples{
lambda(c=1500,f=200000)
}
