% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pil_fb-data.R
\docType{data}
\name{pil_fb}
\alias{pil_fb}
\title{Sardine Fish body Shape}
\format{
data.frame
}
\source{
https://www.fisheries.noaa.gov/data-tools/krm-model
}
\usage{
data(pil_fb)
}
\description{
Example Shapes of the fish body (fb) of a few sticklebacks (stb)
}
\examples{
data(pil_fb)
fb=pil_fb
par(mfrow=c(1,2))
KRMr::shplot(x_fb = fb$x_fb, w_fb = fb$w_fb,
            z_fbU = fb$z_fbU, z_fbL = fb$z_fbL)
}
\references{
NOAA Southwest Fisheries Science Center
}
\keyword{datasets}
