% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pil_sb-data.R
\docType{data}
\name{pil_sb}
\alias{pil_sb}
\title{Sardine Swimbladder Shape}
\format{
data.frame
}
\source{
https://www.fisheries.noaa.gov/data-tools/krm-model
}
\usage{
data(pil_sb)
}
\description{
Example Shapes of the swimbladder (sb) of a few sticklebacks (stb)
}
\examples{
data(pil_sb)
sb=pil_sb
par(mfrow=c(1,2))
KRMr::shplot(x_sb = sb$x_sb, w_sb = sb$w_sb,
            z_sbU = sb$z_sbU, z_sbL = sb$z_sbL)
}
\references{
NOAA Southwest Fisheries Science Center
}
\keyword{datasets}
