% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KRM.R
\name{shplot}
\alias{shplot}
\title{Plots the KRM input shape}
\usage{
shplot(
  x_fb = NULL,
  x_sb = NULL,
  w_fb = NULL,
  w_sb = NULL,
  z_fbU = NULL,
  z_fbL = NULL,
  z_sbU = NULL,
  z_sbL = NULL,
  x_sb2 = NULL,
  w_sb2 = NULL,
  z_sbU2 = NULL,
  z_sbL2 = NULL,
  pa = FALSE
)
}
\arguments{
\item{x_fb}{Position x direction Fish body}

\item{x_sb}{Position x direction Swimbladder}

\item{w_fb}{Width at position x fish body}

\item{w_sb}{Width at position x Swimbladder}

\item{z_fbU}{Height fish body at position x, upper value}

\item{z_fbL}{Height fish body at position x, lower value}

\item{z_sbU}{Height swimbladder at position x, upper value}

\item{z_sbL}{Height swimbladder at position x, lower value}

\item{x_sb2}{Position x direction second Swimbladder}

\item{w_sb2}{Width at position x second Swimbladder}

\item{z_sbU2}{Height second swimbladder at position x, upper value}

\item{z_sbL2}{Height second swimbladder at position x, lower value}
}
\description{
Plots the KRM input shape
}
