% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_fb-data.R
\docType{data}
\name{stb_fb}
\alias{stb_fb}
\title{Stickleback Fish Body Shape}
\format{
data.frame
}
\source{
TBA
}
\usage{
data(stb_fb)
}
\description{
Example Shapes of the fishbody (fb) of a few sticklebacks (stb)
}
\examples{
data(stb_fb)
i = unique(stb_fb$ind)[1]
fb =stb_fb\%>\%filter(ind==i)
par(mfrow=c(1,2))
KRMr::shplot(x_fb = fb$x_fb, w_fb = fb$w_fb,
            z_fbU = fb$z_fbU, z_fbL = fb$z_fbL)
}
\references{
TBA
}
\keyword{datasets}
