% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_sb-data.R
\docType{data}
\name{stb_sb}
\alias{stb_sb}
\title{Stickleback Swimbladder Shape}
\format{
data.frame
}
\source{
TBA
}
\usage{
data(stb_sb)
}
\description{
Example Shapes of the swimbladder (sb) of a few sticklebacks (stb)
}
\examples{
data(stb_sb)
i = unique(stb_sb$ind)[1]
sb =stb_sb\%>\%filter(ind==i)
par(mfrow=c(1,2))
KRMr::shplot(x_sb = sb$x_sb, w_sb = sb$w_sb,
            z_sbU = sb$z_sbU, z_sbL = sb$z_sbL)
}
\references{
TBA
}
\keyword{datasets}
